+++ conn_test.go
@@ -2073,3 +2073,28 @@ func TestConnInitConnInfo(t *testing.T) {
 
 	ensureConnValid(t, conn)
 }
+
+func TestDomainType(t *testing.T) {
+	conn := mustConnect(t, *defaultConnConfig)
+	defer closeConn(t, conn)
+
+	dt, ok := conn.ConnInfo.DataTypeForName("uint64")
+	if !ok {
+		t.Fatal("Expected data type for domain uint64 to be present")
+	}
+	if dt, ok := dt.Value.(*pgtype.Numeric); !ok {
+		t.Fatal("Expected data type value for domain uint64 to be *pgtype.Numeric, but it was %T", dt)
+	}
+
+	var n uint64
+	err := conn.QueryRow("select $1::uint64", uint64(42)).Scan(&n)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if n != 42 {
+		t.Fatalf("Expected n to be 42, but was %v", n)
+	}
+
+	ensureConnValid(t, conn)
+}