+++ pgmock/pgmock.go
@@ -479,6 +479,55 @@ where (
 		SendMessage(&pgproto3.ReadyForQuery{TxStatus: 'I'}),
 	}...)
 
+	steps = append(steps, []Step{
+		ExpectMessage(&pgproto3.Parse{
+			Query: "select t.oid, t.typname, t.typbasetype\nfrom pg_type t\n  join pg_type base_type on t.typbasetype=base_type.oid\nwhere t.typtype = 'd'\n  and base_type.typtype = 'b'",
+		}),
+		ExpectMessage(&pgproto3.Describe{
+			ObjectType: 'S',
+		}),
+		ExpectMessage(&pgproto3.Sync{}),
+		SendMessage(&pgproto3.ParseComplete{}),
+		SendMessage(&pgproto3.ParameterDescription{}),
+		SendMessage(&pgproto3.RowDescription{
+			Fields: []pgproto3.FieldDescription{
+				{Name: "oid",
+					TableOID:             1247,
+					TableAttributeNumber: 65534,
+					DataTypeOID:          26,
+					DataTypeSize:         4,
+					TypeModifier:         4294967295,
+					Format:               0,
+				},
+				{Name: "typname",
+					TableOID:             1247,
+					TableAttributeNumber: 1,
+					DataTypeOID:          19,
+					DataTypeSize:         64,
+					TypeModifier:         4294967295,
+					Format:               0,
+				},
+				{Name: "typbasetype",
+					TableOID:             1247,
+					TableAttributeNumber: 65534,
+					DataTypeOID:          26,
+					DataTypeSize:         4,
+					TypeModifier:         4294967295,
+					Format:               0,
+				},
+			},
+		}),
+		SendMessage(&pgproto3.ReadyForQuery{TxStatus: 'I'}),
+		ExpectMessage(&pgproto3.Bind{
+			ResultFormatCodes: []int16{1, 1, 1},
+		}),
+		ExpectMessage(&pgproto3.Execute{}),
+		ExpectMessage(&pgproto3.Sync{}),
+		SendMessage(&pgproto3.BindComplete{}),
+		SendMessage(&pgproto3.CommandComplete{CommandTag: "SELECT 0"}),
+		SendMessage(&pgproto3.ReadyForQuery{TxStatus: 'I'}),
+	}...)
+
 	return steps
 }