+++ travis/before_script.bash
@@ -9,6 +9,7 @@ then
   # of aclitem formatting. It turns out aclitems cannot contain non-existing users/roles.
   psql -U postgres -c 'create database pgx_test'
   psql -U postgres pgx_test -c 'create extension hstore'
+  psql -U postgres pgx_test -c 'create domain uint64 as numeric(20,0)'
   psql -U postgres -c "create user pgx_ssl SUPERUSER PASSWORD 'secret'"
   psql -U postgres -c "create user pgx_md5 SUPERUSER PASSWORD 'secret'"
   psql -U postgres -c "create user pgx_pw  SUPERUSER PASSWORD 'secret'"