+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Support sslkey, sslcert, and sslrootcert URI params (Sean Chittenden)
 * Allow any scheme in ParseURI (for convenience with cockroachdb) (Sean Chittenden)
+* Add support for domain types
 
 ## Fixes